requirements.txt
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyVirtualDisplay==3.0
requests==2.31.0
rsa==4.9
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
